+++ assets/javascripts/kitten/components/molecules/modal-next/Modal.stories.js
@@ -1,7 +1,7 @@
 import React, { useState } from 'react'
 import { text, boolean, number, select } from '@storybook/addon-knobs'
 import { Modal } from './index'
-import { Button, Text, SaveIcon } from '../../..'
+import { Button, Text, SaveIcon, COLORS } from '../../..'
 
 const paragraphContainer = `
   Sed ut perspiciatis unde omnis iste natus error sit voluptatem
@@ -159,15 +159,6 @@ export const FullSize = () => (
   </Modal>
 )
 
-const OrionHeaderTitle = () => {
-  return (
-    <p>
-      <strong>Hello world !</strong>
-      <br />
-      Hey
-    </p>
-  )
-}
 const OrionHeaderActions = ({ close }) => (
   <>
     <Button
@@ -200,8 +191,13 @@ export const ComplexWithOrion = () => (
   <Modal
     trigger={<Button modifier="helium">Open</Button>}
     variant="orion"
-    headerTitle={<OrionHeaderTitle />}
+    headerTitle={<Text weight="bold" size="giant">Modal title</Text>}
     headerActions={OrionHeaderActions}
+    headerMessage={
+      <Text size="micro" weight="light" cssColor={COLORS.grey1}>
+        {text('headerMessage', '')}
+      </Text>
+    }
     size={select('Size', ['regular', 'big', 'huge', 'giant'], 'giant')}
   >
     {() => (